Move trait helpers to AbstractCli class
For now all of them seem not to be reused, except the cliError helper. So it makes sense to have them in the abstract class.
